자유게시판 이름 변경(freeboard > freeboardlist, freeboarddetail > freeboardwrite. postdetail > freeboarddetail
